pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "challenge-validator"
version = "0.1.0"
description = "Test your bot-detection / CAPTCHA SDK integration. 22 tests, 7 categories, fix guides included."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Kaustubh Phatak"}]
keywords = ["captcha", "bot-detection", "waf", "security", "testing", "challenge", "turnstile", "recaptcha"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
pretty = ["rich>=13.0"]

[project.scripts]
challenge-validator = "challenge_validator.cli:main"

[project.urls]
Homepage = "https://github.com/kphatak001/challenge-validator"
Documentation = "https://github.com/kphatak001/challenge-validator/tree/main/docs"
Issues = "https://github.com/kphatak001/challenge-validator/issues"

[tool.setuptools.packages.find]
include = ["challenge_validator*"]

[tool.setuptools.package-data]
challenge_validator = ["profiles/*.yaml"]
